+++ packages/form-render/src/resolver.ts
@@ -1,0 +1,15 @@
+export function FormRenderResolver() {
+  return {
+    type: 'component' as const,
+    resolve: (name: string) => {
+      const packageName = '@gopowerteam/form-render'
+      if (name === 'FormRender' || name === 'form-render') {
+        return {
+          name: 'FormRender',
+          from: packageName,
+          sideEffects: [`${packageName}/dist/style.css`],
+        }
+      }
+    },
+  }
+}