+++ packages/form-render/vite.config.ts
@@ -14,7 +14,7 @@ export default defineConfig({
     lib: {
       entry: [
         resolve(__dirname, 'src/index.ts'),
-        // resolve(__dirname, 'resolver.ts'),
+        resolve(__dirname, 'src/resolver.ts'),
       ],
       name: 'bundle',
       fileName: (format, name) => `${format}/${name}.${format === 'es' ? 'mjs' : 'cjs'}`,